Removes extra button component
